feat: transfer log to sentinel fn to sentinel integration
