+++ app/tests/commands/test_utils.py
@@ -12,22 +12,6 @@ def test_log_ops_message():
     )
 
 
-@patch("commands.utils.send_event")
-def test_log_to_sentinel(send_event_mock):
-    utils.log_to_sentinel("foo", {"bar": "baz"})
-    send_event_mock.assert_called_with({"event": "foo", "message": {"bar": "baz"}})
-
-
-@patch("commands.utils.send_event")
-@patch("commands.utils.logging")
-def test_log_to_sentinel_logs_error(logging_mock, send_event_mock):
-    send_event_mock.return_value = False
-    utils.log_to_sentinel("foo", {"bar": "baz"})
-    send_event_mock.assert_called_with({"event": "foo", "message": {"bar": "baz"}})
-    logging_mock.error.assert_called_with(
-        "Sentinel event failed: {'event': 'foo', 'message': {'bar': 'baz'}}"
-    )
-
 
 def test_get_user_locale_supported_locale():
     client = MagicMock()
